Combine intensity and ret/azim raytrace functions
The Jones matrices corresponding to the polarizers and analyzers are now no longer treated with the rest of the material voxels. They can be applied separately so that the ray tracing through the voxels only needs to be done once when acquiring multiple intensity images with various polarizer settings.

The ray_trace_through_volume function now returns a list. Some of the intensity function have not been implented in pytorch. Hence, intensity testing with pytorch has not been performed. For the retardance and azimuth images, the new function calc_cummulative_JM_lenslet_numpy can also be used to avoid needed to iterated over all the lenslet pixels.

See forward_intensity.py for the forward simulation that generates the intensity images.

Removed files that are not in use and significantly cleaned many of the scripts. All tests pass.