Add plotting of estimated volume
There was an error with the nticks of the plotly. So I created an option
to disable the nticks parameter of the scene layout.
For the plotting, the optic axis should be normalized first.

I added download buttons for the estimated images and button.
Use of the gpu should be made optional, even if it is available.
More operations in the reconstruction process should be turned into
their own function.

All tests pass.