Jones calculus azimuth rotation, big commit
I created modular versions of the main forward and reconstuction scripts. The common functions in the scripts should be moved to a new script. I replaced the colormap of the orientation images with a cyclic colormap.

Birefringence implementations: I added a normalized option. In the Jones calculus, the projection of the ray onto the optic axis is normalized. I improved the method of initializing volumes. I redefined how the ellisoid is created. It is not the substraction of two ellisoids of slightly different radii. I added a Jones calculus pytorch implementation to account for the affect of negative birefringence on the azimuth image. The slow axis should be rotated 90 degrees when the sign of the birefringence is flipped.

Pytest updates: I created a test to verify the physics affects of having opposite sign on the birefringence. I added basic BirefringentVolume tests. I removed the depreciated torch.set_default_tensor_type() function. To check the the gradients were not broken, I increased the volume size to make it be more likely that the volume contained nonzero birefringence and verified that the gradients are not None instead of not being zero. The new ellipsoid volume was causing the test to otherwise fail. A test should be created to verify that the optimizer does not update the values to be NaNs, and that is why the updated volume is not the same as the initial volume.

All tests pass.