Add attributes to Reconstructor
I made the Reconstructor class more modular. I wrote visualization functions that get imported.

In the BirefringentVolume class, I created a method to load a volume from a h5 file. It needs improvement. The idea is that the method will serve an alternative to the init_from_file method that does not depend on optical_info.

In plotting_tools.py, I removed more of the waveblocks leftover dependency.

All tests pass.