Optimize birefringence subsets
The forward model can now be based on a new Delta_n variable. This variable can be a torch concatenation of two torch arrays, one with gradients and one not. A similar approach should be done for optic_axis. To use, change OPTIMIZING_MODE to True at the top of birefringence_implementations.py.

All tests pass.