New masking methodology and ability to optimize subset
Mask creation and ray filtering improvements: Now voxels are only considered in the mask if they contribute to at least one nonzero pixel in the retardance image.

If the birefringent volume has the attribute indices_active as not None, then the smaller set of property arrays are assumed to be in use.