Optimize over only 2 optic axis components
Using only 2 of the 3 components seems to increase the reconstruction iterations by about 40%. I also sped up the ret and azim calculations by only computing the ret and azim when the birefringence of the voxel is nonzero.

Instead of using a dictionary to map from the subset of voxels being optimized to their spatial coordinates, I switched to using a tensor. This is much faster to retrieve values out of. This sped up the gathering parameters time by 20%, which was surprisingly a comparable time to the Jones matrix multiplication time.