Forward model arctan swap ordering
Swapped the order of the input to arctan2 when computing the azimuth of a Jones matrix. This forward model now aligns with the LC-PolScope.

Storing the collision indices as a tensor instead of a list for computational efficiency purposes when ray tracing through the volume. Huge speed improvement.

Ray geometry can now be saved to then be loaded at the beginning of a reconstruction process. This decreases the time before the 1st iteration by 10x.