SU2 retardance calculation method
By exploting the SU2 property of the jones matrix, I was able to improve the retardance calculation speed by about 100x.

When comparing these two retardance calculation methods, the test comparison tolarance level needed to be as high as 5e-3. This could be looked into further.

I adjusted the assert statement for the test that counted the indices raytraced through volume because the output was inconsistent. A similar inconsistency was mentioned here https://github.com/PolarizedLightFieldMicroscopy/GeoBirT/issues/90#issuecomment-2087394284