Radiometry mask
Mask from radiometry image:
I developed the ability to mask the valid rays based on the radiometry images. Furthermore, the voxels associated with the valid rays can be filtered when creating a voxel mask.
The masking is preformed one the MLA geometry is prepared. Thus, the masking is done with tensor and without looping across the lenslets, giving a huge speed improvement.

Ray saving:
With the addition of the radiometry filtering, the ray saving is done after all the MLA geometry is created, but before the mask creation.

Jones calculus:
The jones matrix formulation is simplified and made much faster. The inputs to arctan are swapped. This fixes the 45 degree axis on the single voxel azimuth image.
If both inputs to arctan are close to zero, the azimuth is set to be zero instead of pi/2. This primarily affects the background of the azimuth image.

Optimization:
Created a scheduler to decrease the learning rate if the loss has not decreased enough recently. Optionally, the gradients can be clipped if their norm is too high.
Warmup iterations are included to avoid giant changes particularly in the 2nd iteration.

Other:
- Created function to be able to go back to a square light field image from the 1D tensor version.
- Large unnessary ray attributes are deleted from memory, particularly before saving the rays.