Extract voxel properties outside of Jones matrix multiplication loop
Now the pytorch simulations are faster, particularly on the GPU.
The matrix multiplication still need to be done in a loop because it is not a commutative operation.
Updated time stamp labels now that a loop is removed.
Adjusting jones matrix function to account for the new dimensions of tensors.
Removed duplicate numpy array saving function.